app/requirements.txt
opentelemetry-api
opentelemetry-sdk
# opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-jaeger